refactor: moved the serializer/deserializer to a dedicated folder
