+++ src/lib/MMKVStorage/encoding/deserializers/date.ts
@@ -1,0 +1,7 @@
+import { Deserializer, SerializedObject } from '../types';
+
+export const DateDeserializer: Deserializer = {
+  decode(serialized: SerializedObject): Date {
+    return new Date(serialized.value);
+  },
+};
