+++ src/lib/MMKVStorage/encoding/deserializers/uint8array.ts
@@ -1,0 +1,8 @@
+import { fromHex } from '@cosmjs/encoding';
+import { Deserializer, SerializedObject } from '../types';
+
+export const Uint8ArrayDeserializer: Deserializer = {
+  decode(serialized: SerializedObject): Uint8Array {
+    return fromHex(serialized.value);
+  },
+};
