+++ src/lib/MMKVStorage/encoding/serializers/index.ts
@@ -1,0 +1,7 @@
+import { Serializer } from '../types';
+import { DateSerializer } from './date';
+import { Uint8ArraySerializer } from './uint8array';
+
+const Serializers: Serializer[] = [DateSerializer, Uint8ArraySerializer];
+
+export default Serializers;
